Finished transition of code to single model.
The purpose of this transition was to get rid of the dupplicate methods
in NoteModel and NotesListViewModel. Now NotesListViewModel is just what
it is supposed to be - a view that now how to display a hierarchy of
notes starting from specified note (root note). The transition
simplified the code and appears to be working good.