Changed the way notes are deleted and fixed undo.
* Added `deleted` property to the Note model. Instead of actually
deleting a note it is marked with `deleted` flag. Models with this flag
set to true are not retrieved by default. Trying to `delete` a note that
is already deleted will do nothing.
* Added `restore` method to restore previously deleted notes. Although
it is not enforced by the server-side, notes should be restored in the
order opposite to the one they were deleted in.
* It will be possible in the future to retrieve deleted notes as well.